feat(US-07): Função de pegar o histórico geral e histórico de tarefa específica

+++ task-service/src/main/java/com/kernelpanic/task_service/repositorios/HistoricoTarefaRepositorio.java
@@ -10,7 +10,7 @@ import com.kernelpanic.task_service.entidades.HistoricoTarefa;
 @Repository
 public interface HistoricoTarefaRepositorio extends JpaRepository<HistoricoTarefa, Long> {
 
-    List<HistoricoTarefa> findByTarefaId(Long tarefaId);
+    List<HistoricoTarefa> findByTarefaId(Integer tarefaId);
     List<HistoricoTarefa> findByTarefaIdOrderByDataEventoAsc(Integer tarefaId);
 
 }
